pyproject.toml
[project]
name = "crawl"
version = "0.1.0"
description = ""
authors = [
    {name = "sungminna",email = "sungmin.na330@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest-playwright (>=0.7.0,<0.8.0)",
    "playwright (>=1.50.0,<2.0.0)",
    "setuptools (>=75.8.1,<76.0.0)",
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "html2text (>=2024.2.26,<2025.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "openai (>=1.65.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
